Moved reorder metric to own component
