+++ src/frontend/src/pages/Metrics.tsx
@@ -6,6 +6,7 @@ import { ArrowLeft } from "lucide-react";
 import { AutoMetricOptionsResponse, Metric, ModalMode } from "./MetricPageTypes";
 import { headerCellStyle, metricCellStyle } from "../components/CellComponents";
 import { AddMetricModal } from "../components/AddMetricModal";
+import { ReorderMetricsModal } from "../components/ReorderMetricsModal";
 import ExpandableText from "../components/ExpandableText";
 const MetricsPage: React.FC = () => {
   const [loading, setLoading] = useState(true);
@@ -912,156 +913,17 @@ const MetricsPage: React.FC = () => {
           
         </div>
 
-        {reorderModalOpen && (
-          <div
-            className="dx-backdrop"
-            role="dialog"
-            aria-modal="true"
-            onMouseDown={(e) => {
-              if (e.target === e.currentTarget) closeReorderModal();
-            }}
-            style={{
-              display: "flex",
-              alignItems: "center",
-              justifyContent: "center",
-              backgroundColor: "rgba(0, 0, 0, 0.6)",
-              zIndex: 9998,
-            }}
-          >
-            <div
-              className="dx-card"
-              onMouseDown={(e) => e.stopPropagation()}
-              style={{
-                width: "min(760px, 92vw)",
-                maxHeight: "85vh",
-                overflow: "auto",
-                padding: 18,
-                position: "relative",
-                background: "rgba(18, 18, 26, 0.98)",
-                border: "1px solid rgba(255,255,255,0.08)",
-                borderRadius: 16,
-                boxShadow: "0px 10px 25px rgba(0, 0, 0, 0.2)",
-              }}
-            >
-              <div
-                style={{
-                  display: "flex",
-                  alignItems: "center",
-                  justifyContent: "space-between",
-                  marginBottom: 14,
-                  gap: 12,
-                  minWidth: 0,
-                }}
-              >
-                <div style={{ fontSize: "1.15rem", fontWeight: 700, color: "var(--accent)" }}>
-                  Reorder Metrics by Category
-                </div>
-
-                <button
-                  className="dx-btn dx-btn-outline"
-                  onClick={closeReorderModal}
-                  aria-label="Close"
-                  style={{ padding: "6px 10px" }}
-                >
-                  ✕
-                </button>
-              </div>
-
-              <div style={{ display: "grid", gap: 12 }}>
-                <div style={{ display: "flex", flexDirection: "column", gap: 6 }}>
-                  <label style={{ opacity: 0.85 }}>Category</label>
-                  <select
-                    className="dx-input"
-                    value={reorderCategory}
-                    onChange={(e) => setReorderCategory(e.target.value)}
-                  >
-                    {categoryOrder.map((category) => (
-                      <option key={category} value={category}>
-                        {category}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-
-                <div
-                  style={{
-                    maxHeight: "52vh",
-                    overflow: "auto",
-                    padding: 12,
-                    borderRadius: 14,
-                    border: "1px solid rgba(255,255,255,0.08)",
-                    background: "rgba(255,255,255,0.03)",
-                  }}
-                >
-                  {(categoryMetricOrder[reorderCategory] || []).length > 0 ? (
-                    (categoryMetricOrder[reorderCategory] || []).map((metricId, index) => {
-                      const metric = metricsById.get(metricId);
-                      if (!metric) return null;
-                      const categoryIds = categoryMetricOrder[reorderCategory] || [];
-                      return (
-                        <div
-                          key={metricId}
-                          style={{
-                            display: "flex",
-                            alignItems: "center",
-                            justifyContent: "space-between",
-                            gap: 10,
-                            padding: "10px 0",
-                            borderBottom: index < categoryIds.length - 1 ? "1px solid rgba(255,255,255,0.08)" : undefined,
-                          }}
-                        >
-                          <div
-                            style={{
-                              minWidth: 0,
-                              overflow: "hidden",
-                              textOverflow: "ellipsis",
-                              whiteSpace: "nowrap",
-                              color: "rgba(255,255,255,0.9)",
-                            }}
-                            title={metric.metric_name}
-                          >
-                            {metric.metric_name}
-                          </div>
-                          <div style={{ display: "flex", gap: 6 }}>
-                            <button
-                              type="button"
-                              className="dx-btn dx-btn-outline"
-                              disabled={index === 0}
-                              onClick={() => moveMetricInCategory(metricId, "up")}
-                              style={{ padding: "5px 8px", fontSize: 12 }}
-                            >
-                              ↑
-                            </button>
-                            <button
-                              type="button"
-                              className="dx-btn dx-btn-outline"
-                              disabled={index === categoryIds.length - 1}
-                              onClick={() => moveMetricInCategory(metricId, "down")}
-                              style={{ padding: "5px 8px", fontSize: 12 }}
-                            >
-                              ↓
-                            </button>
-                          </div>
-                        </div>
-                      );
-                    })
-                  ) : (
-                    <div style={{ opacity: 0.7, padding: 12 }}>No metrics found for this category.</div>
-                  )}
-                </div>
-
-                <div style={{ display: "flex", justifyContent: "flex-end", gap: 10 }}>
-                  <button className="dx-btn dx-btn-outline" onClick={closeReorderModal}>
-                    Cancel
-                  </button>
-                  <button className="dx-btn dx-btn-primary" onClick={saveReorder}>
-                    Save Order
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
+        <ReorderMetricsModal
+          isOpen={reorderModalOpen}
+          categoryOrder={categoryOrder}
+          reorderCategory={reorderCategory}
+          categoryMetricOrder={categoryMetricOrder}
+          metricsById={metricsById}
+          onCategoryChange={setReorderCategory}
+          onMoveMetric={moveMetricInCategory}
+          onClose={closeReorderModal}
+          onSave={saveReorder}
+        />
 
         <AddMetricModal
           isOpen={isModalOpen}
